backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-chess==1.999
